% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saxtonrawls.R
\name{saxtonrawls}
\alias{saxtonrawls}
\title{Use particle size analysis data to determine
Soil Water Characteristics based on Saxton and Rawls (2006)
Available at: https://doi.org/10.2136/sssaj2005.0117}
\usage{
saxtonrawls(sand, clay, OM, DF = 1, Rw = 0, Rv = NULL)
}
\arguments{
\item{sand}{numeric, sand (0.5 - 2 mm) content in decimal format (e.g. 0.25)}

\item{clay}{numeric, clay (<0.002 mm) content in decimal format (e.g. 0.25)}

\item{OM}{numeric, Organic matter content in percentage (e.g., 2.5)}

\item{DF}{numeric, density factor, default value 1}

\item{Rw}{numeric, Volume fraction of gravel in decimal format, g/cm3, default is 0.}

\item{Rv}{numeric, Weight fraction of gravel in decimal format, g/g, default is NULL.}
}
\value{
matrix
}
\description{
Use particle size analysis data to determine
Soil Water Characteristics based on Saxton and Rawls (2006)
Available at: https://doi.org/10.2136/sssaj2005.0117
}
\examples{
#Determine soil water characteristics for a single observation
saxtonrawls(sand=0.45, clay=0.35, OM=4, DF=1, Rw=0.10)


#Determine soil water characteristics for multiple observations
dat<- data.frame(sand=c(20,40,80), clay= c(65,30,10), OM=c(2,2,2), DF=c(1,1,1), Rw=c(0.1, 0.1, 0.1))
mapply(saxtonrawls,sand=dat$sand, clay=dat$clay, OM=dat$OM, DF=dat$DF, Rw=dat$Rw)


}
